% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fitted_probs}
\alias{fitted_probs}
\title{Compute fitted probabilities for an object of class "pmCDM".}
\usage{
fitted_probs(mod, U, control = list(), ...)
}
\arguments{
\item{mod}{\code{pmCDM} object.}

\item{U}{Points for latent attributes on which compute the fitted probabilities}

\item{control}{Additional values for control}
}
\value{
A matrix with fitted probabilities (one entry per row in \code{U}):
\itemize{
\item \code{PI}: Matrix of fitted probabilities.
}
}
\description{
Compute fitted probabilities for an object of class "pmCDM".
}
\details{
Test
}
\author{
Camilo Cárdenas-Hurtado (\email{c.a.cardenas-hurtado@lse.ac.uk}).
}
